refactor: 불필요한 PaymentInfo를 제거
