+++ src/test/java/roomescape/payment/application/PaymentServiceTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.Mockito.when;
 
-import java.time.LocalDate;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -17,9 +16,8 @@ import roomescape.payment.application.dto.TossConfirmRequest;
 import roomescape.payment.application.dto.TossConfirmResponse;
 import roomescape.payment.application.dto.TossConfirmResponse.EasyPayInfo;
 import roomescape.payment.domain.Payment;
-import roomescape.payment.domain.PaymentInfo;
+import roomescape.payment.domain.PaymentGateway;
 import roomescape.payment.infra.TossPaymentGatewayClient;
-import roomescape.reservation.application.dto.MemberReservationRequest;
 
 @ActiveProfiles("test")
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
@@ -35,9 +33,7 @@ class PaymentServiceTest {
     @Test
     void 결제_저장_테스트() {
         // given
-        PaymentInfo paymentInfo = new PaymentInfo("key", "orderId", 1000L);
-        PaymentRequest paymentRequest = new PaymentRequest(paymentInfo.getPaymentKey(),
-            paymentInfo.getOrderId(), paymentInfo.getAmount());
+        PaymentRequest paymentRequest = new PaymentRequest("key", "orderId", 1000L);
         TossConfirmRequest tossConfirmRequest = new TossConfirmRequest(paymentRequest.paymentKey(),
             paymentRequest.orderId(), paymentRequest.amount());
 
@@ -54,17 +50,10 @@ class PaymentServiceTest {
         // then
         assertAll(
             () -> assertThat(payment.getId()).isNotNull(),
-            () -> assertThat(payment.getPaymentInfo().equals(paymentInfo))
+            () -> assertThat(payment.getPaymentKey().equals(paymentRequest.paymentKey())),
+            () -> assertThat(payment.getOrderId().equals(paymentRequest.orderId())),
+            () -> assertThat(payment.getAmount()).isEqualTo(paymentRequest.amount()),
+            () -> assertThat(payment.getPaymentGateway()).isEqualTo(PaymentGateway.TOSS_PAYMENTS)
         );
-    }
-
-    private MemberReservationRequest createRequest(
-        LocalDate now,
-        Long timeId,
-        Long themeId,
-        PaymentInfo paymentInfo
-    ) {
-        return new MemberReservationRequest(now, timeId, themeId, paymentInfo.getPaymentKey(),
-            paymentInfo.getOrderId(), paymentInfo.getAmount());
     }
 }
